refactor(noname): extract review guardrail helpers
