+++ docs/项目文档/03-规划/01-现状评估.md
@@ -15,9 +15,17 @@
 
 ## 3. 当前主要风险
 
-### 3.1 命令主链压力过大
+### 3.1 命令主链压力仍需继续收口
 
-`tauri_commands.rs` 仍然承担了过多编排工作，继续向里面叠逻辑会提高维护风险和冲突概率。
+`tauri_commands.rs` 仍然承担大量主游戏编排工作，继续向里面叠逻辑会提高维护风险和冲突概率。
+
+当前 `NoName reviewed apply` 链路已经完成第三轮减压：
+
+- reviewed apply request 构造、快照校验与显式 apply 已下沉到 `noname_apply.rs`
+- trace 获取 / plot state 更新 / trace 回写已复用命令层共用壳层
+- human review intent 与 second guardrail 决策记录已下沉到 `noname_apply.rs`
+
+后续风险重点不再是 reviewed apply 本身，而是其他低风险 apply 与剧情生成主链是否继续膨胀。
 
 ### 3.2 前后端 mock 语义漂移
