+++ docs/项目文档/03-规划/02-后续路线图.md
@@ -53,10 +53,12 @@
 
 当前状态：
 
-- 已完成第二轮下沉
+- 已完成第三轮下沉
 - reviewed apply 的 request builder 与 plotText manual apply helper 已进入 `noname_apply.rs`
 - `tauri_commands.rs` 在 reviewed apply 路径上已减少内联快照构造与包装逻辑
 - reviewed apply 命令已开始复用共享的 trace 读取、plot state 更新和 trace 回写壳层
+- human review intent 与 second guardrail 决策记录已进入 `noname_apply.rs`
+- 下一步若继续减压，应优先评估低风险 apply 或剧情生成主链中仍重复的编排，而不是扩大新的 apply scope
 
 ## P2：收口后继续推进
