+++ docs/项目文档/03-规划/04-未来两周开发排期.md
@@ -91,10 +91,11 @@
 
 当前状态：
 
-- 第二轮已完成
+- 第三轮已完成
 - `NoNameReviewedApplyRequest` 构造与 plotText manual apply helper 已从命令层下沉到 `noname_apply.rs`
 - trace 获取 / result 回写的共用壳层已开始抽出并复用到两个 reviewed apply 命令
-- 下一轮可继续观察 review / second guardrail 命令是否也值得抽出同类 runtime helper
+- human review intent 与 second guardrail 决策记录已从命令层下沉到 `noname_apply.rs`
+- 下一轮若继续 A2，应只评估剩余低风险 apply / 生成主链重复编排，不新增 apply scope，不扩张 second guardrail 语义
 
 ### 任务 A3：建立 `NoName` 主链变更的文档闭环
