+++ src-tauri/src/tauri_commands.rs
@@ -16,7 +16,8 @@ use crate::llm_service::{LLMConfig, LLMRequest, LLMService};
 use crate::memory_layers::{ChapterSummary, MemoryEntry, WorldFact};
 use crate::noname_apply::{
     apply_manual_plot_text_hint_to_plot_state, apply_reviewed_output_to_plot_state,
-    build_reviewed_apply_request,
+    build_reviewed_apply_request, record_human_review_apply_intent,
+    record_second_guardrail_decision,
 };
 use crate::noname_config::NoNameConfig;
 use crate::noname_context_builder::build_context_packet;
@@ -243,7 +244,10 @@ fn apply_noname_outcome_with_engine<F>(
     apply: F,
 ) -> Result<NoNameManualPlotTextApplyResult, String>
 where
-    F: FnOnce(NoNameTrace, PlotState) -> Result<crate::noname_apply::NoNameReviewedApplyOutcome, String>,
+    F: FnOnce(
+        NoNameTrace,
+        PlotState,
+    ) -> Result<crate::noname_apply::NoNameReviewedApplyOutcome, String>,
 {
     let trace = load_noname_trace(trace_id)?;
     let result = {
@@ -464,306 +468,6 @@ fn record_noname_apply_plan_and_skip_execution(
     if let NoNameApplyTargetDecision::Skip { outcome, note } = &plan.decision {
         trace.record_apply_execution(plan.target, *outcome, Some(note.clone()));
     }
-}
-
-fn next_noname_apply_plan_order(trace: &NoNameTrace) -> u32 {
-    trace
-        .apply_plan_log
-        .iter()
-        .map(|item| item.order)
-        .max()
-        .unwrap_or_default()
-        + 1
-}
-
-fn find_review_proposal<'a>(
-    trace: &'a NoNameTrace,
-    request_id: &str,
-) -> Option<&'a crate::noname_types::NoNameProposal> {
-    trace
-        .proposals
-        .iter()
-        .rev()
-        .find(|proposal| request_id.contains(&proposal.proposal_id))
-        .or_else(|| trace.proposals.last())
-}
-
-fn record_noname_human_review_apply_intent(
-    trace: &mut NoNameTrace,
-    request_id: &str,
-    decision: NoNameHumanReviewDecision,
-    safe_apply_scope: Option<NoNameApplyScope>,
-) {
-    let target = safe_apply_scope
-        .map(|scope| scope.as_str())
-        .unwrap_or("controlled_output");
-    let order = next_noname_apply_plan_order(trace);
-    let priority = safe_apply_scope.map(priority_for_apply_scope).unwrap_or(10) + 25;
-
-    match decision {
-        NoNameHumanReviewDecision::Pending => {
-            trace.record_apply_plan(
-                order,
-                target,
-                "review_intent_pending",
-                priority,
-                Some("人工复核已重置为待确认，未进入二次 guardrail".to_string()),
-            );
-            trace.record_apply_execution(
-                target,
-                "human_review_pending",
-                Some("等待人工确认，不触发高层 apply".to_string()),
-            );
-            trace.record_proposal_transition(format!("{}:apply_intent:pending", request_id));
-        }
-        NoNameHumanReviewDecision::RejectedForHigherApply => {
-            trace.record_apply_plan(
-                order,
-                target,
-                "reject",
-                priority,
-                Some("人工复核拒绝进入高层 apply，保持安全边界".to_string()),
-            );
-            trace.record_apply_execution(
-                target,
-                "rejected_by_human_review",
-                Some("开发者选择暂不应用，未触发二次 guardrail".to_string()),
-            );
-            trace.record_proposal_transition(format!(
-                "{}:apply_intent:rejected_by_human_review",
-                request_id
-            ));
-        }
-        NoNameHumanReviewDecision::ApprovedForHigherApply => {
-            let rejection_reason = if trace.mode != NoNameMode::Assisted {
-                Some("当前 trace 不是 assisted 模式，拒绝进入高层 apply intent".to_string())
-            } else if safe_apply_scope != Some(NoNameApplyScope::PlotTextHint) {
-                Some(format!(
-                    "当前仅允许 plot_text_hint 进入人工确认后的二次 guardrail，实际作用域为 {}",
-                    target
-                ))
-            } else {
-                match find_review_proposal(trace, request_id) {
-                    Some(proposal)
-                        if proposal.status
-                            == crate::noname_types::NoNameProposalStatus::Applied
-                            && proposal.applyable
-                            && target_segment_supports_apply_scope(
-                                proposal.target_segment,
-                                NoNameApplyScope::PlotTextHint,
-                            ) =>
-                    {
-                        None
-                    }
-                    Some(proposal)
-                        if proposal.status
-                            != crate::noname_types::NoNameProposalStatus::Applied
-                            || !proposal.applyable =>
-                    {
-                        Some(format!(
-                            "proposal 状态为 {} / applyable={}，拒绝进入高层 apply intent",
-                            proposal.status.as_str(),
-                            proposal.applyable
-                        ))
-                    }
-                    Some(proposal) => Some(format!(
-                        "target_segment={} 不支持 plot_text_hint 二次 apply",
-                        proposal.target_segment.as_str()
-                    )),
-                    None => {
-                        Some("未找到 review 对应 proposal，拒绝进入高层 apply intent".to_string())
-                    }
-                }
-            };
-
-            if let Some(reason) = rejection_reason {
-                trace.record_apply_plan(
-                    order,
-                    target,
-                    "second_guardrail_reject",
-                    priority,
-                    Some(reason.clone()),
-                );
-                trace.record_apply_execution(target, "second_guardrail_rejected", Some(reason));
-                trace.record_proposal_transition(format!(
-                    "{}:apply_intent:second_guardrail_rejected",
-                    request_id
-                ));
-                return;
-            }
-
-            trace.record_apply_plan(
-                order,
-                target,
-                "review_intent_ready",
-                priority,
-                Some(
-                    "人工确认已通过，已排入二次 guardrail / apply planner，未写入正文".to_string(),
-                ),
-            );
-            trace.record_apply_execution(
-                target,
-                "awaiting_second_guardrail",
-                Some("高层 apply intent 已记录，等待后续二次 guardrail 决策".to_string()),
-            );
-            trace.record_proposal_transition(format!(
-                "{}:apply_intent:awaiting_second_guardrail",
-                request_id
-            ));
-        }
-    }
-}
-
-fn review_is_waiting_for_second_guardrail(trace: &NoNameTrace, request_id: &str) -> bool {
-    let transition = format!("{}:apply_intent:awaiting_second_guardrail", request_id);
-    trace
-        .proposal_transition_log
-        .iter()
-        .any(|entry| entry == &transition)
-        || trace.apply_execution_log.iter().any(|entry| {
-            entry.outcome == "awaiting_second_guardrail"
-                && entry.target == NoNameApplyScope::PlotTextHint.as_str()
-        })
-}
-
-fn second_guardrail_revalidation_reason(
-    trace: &NoNameTrace,
-    request_id: &str,
-    safe_apply_scope: Option<NoNameApplyScope>,
-) -> Option<String> {
-    if trace.mode != NoNameMode::Assisted {
-        return Some("当前 trace 不是 assisted 模式，二次 guardrail 拒绝".to_string());
-    }
-    if safe_apply_scope != Some(NoNameApplyScope::PlotTextHint) {
-        return Some("当前二次 guardrail 只允许处理 plot_text_hint".to_string());
-    }
-    if !review_is_waiting_for_second_guardrail(trace, request_id) {
-        return Some("review 尚未进入 awaiting_second_guardrail 队列".to_string());
-    }
-
-    match find_review_proposal(trace, request_id) {
-        Some(proposal)
-            if proposal.status == crate::noname_types::NoNameProposalStatus::Applied
-                && proposal.applyable
-                && target_segment_supports_apply_scope(
-                    proposal.target_segment,
-                    NoNameApplyScope::PlotTextHint,
-                ) =>
-        {
-            None
-        }
-        Some(proposal)
-            if proposal.status != crate::noname_types::NoNameProposalStatus::Applied
-                || !proposal.applyable =>
-        {
-            Some(format!(
-                "proposal 状态为 {} / applyable={}，二次 guardrail 拒绝",
-                proposal.status.as_str(),
-                proposal.applyable
-            ))
-        }
-        Some(proposal) => Some(format!(
-            "target_segment={} 不支持 plot_text_hint 二次 guardrail",
-            proposal.target_segment.as_str()
-        )),
-        None => Some("未找到 review 对应 proposal，二次 guardrail 拒绝".to_string()),
-    }
-}
-
-fn record_noname_second_guardrail_decision(
-    trace: &mut NoNameTrace,
-    request_id: &str,
-    decision: NoNameSecondGuardrailDecision,
-    safe_apply_scope: Option<NoNameApplyScope>,
-) -> Result<(), String> {
-    let target = safe_apply_scope
-        .map(|scope| scope.as_str())
-        .unwrap_or("controlled_output");
-    let order = next_noname_apply_plan_order(trace);
-    let priority = safe_apply_scope.map(priority_for_apply_scope).unwrap_or(10) + 50;
-    let revalidation_reason =
-        second_guardrail_revalidation_reason(trace, request_id, safe_apply_scope);
-
-    if let Some(reason) = revalidation_reason {
-        trace.record_apply_plan(
-            order,
-            target,
-            "second_guardrail_reject",
-            priority,
-            Some(reason.clone()),
-        );
-        trace.record_apply_execution(target, "second_guardrail_rejected", Some(reason.clone()));
-        trace.record_proposal_transition(format!("{}:second_guardrail:reject", request_id));
-        trace.set_apply_result(true, "second_guardrail_reject", Some(reason));
-        return Ok(());
-    }
-
-    match decision {
-        NoNameSecondGuardrailDecision::Allow => {
-            trace.record_apply_plan(
-                order,
-                target,
-                "second_guardrail_allow",
-                priority,
-                Some("二次 guardrail 允许进入后续人工 apply 命令；当前不写正文".to_string()),
-            );
-            trace.record_apply_execution(
-                target,
-                "second_guardrail_allowed",
-                Some("已允许进入下一步人工 apply，但未改写剧情正文".to_string()),
-            );
-            trace.record_proposal_transition(format!("{}:second_guardrail:allow", request_id));
-            trace.set_apply_result(
-                true,
-                "second_guardrail_allow",
-                Some("二次 guardrail 已允许；等待显式人工 apply 命令".to_string()),
-            );
-        }
-        NoNameSecondGuardrailDecision::Reject => {
-            trace.record_apply_plan(
-                order,
-                target,
-                "second_guardrail_reject",
-                priority,
-                Some("二次 guardrail 人工拒绝高层 apply".to_string()),
-            );
-            trace.record_apply_execution(
-                target,
-                "second_guardrail_rejected",
-                Some("二次 guardrail 决策为拒绝，未改写剧情正文".to_string()),
-            );
-            trace.record_proposal_transition(format!("{}:second_guardrail:reject", request_id));
-            trace.set_apply_result(
-                true,
-                "second_guardrail_reject",
-                Some("二次 guardrail 已拒绝高层 apply".to_string()),
-            );
-        }
-        NoNameSecondGuardrailDecision::Fallback => {
-            trace.push_stage(crate::noname_types::NoNameTraceStage::ApplyFallback);
-            trace.fallback_used = true;
-            trace.record_apply_plan(
-                order,
-                target,
-                "second_guardrail_fallback",
-                priority,
-                Some("二次 guardrail 要求回退经典链路".to_string()),
-            );
-            trace.record_apply_execution(
-                target,
-                "second_guardrail_fallback",
-                Some("已记录回退决策，未改写剧情正文".to_string()),
-            );
-            trace.record_proposal_transition(format!("{}:second_guardrail:fallback", request_id));
-            trace.set_apply_result(
-                true,
-                "second_guardrail_fallback",
-                Some("二次 guardrail 决策为 fallback，继续依赖经典链路".to_string()),
-            );
-        }
-    }
-
-    Ok(())
 }
 
 fn apply_noname_plot_text_hint(
@@ -4433,7 +4137,7 @@ pub async fn mark_noname_controlled_output_review(
     }
 
     trace.record_proposal_transition(format!("{}:human_review:{}", request_id, decision.as_str()));
-    record_noname_human_review_apply_intent(&mut trace, &request_id, decision, safe_apply_scope);
+    record_human_review_apply_intent(&mut trace, &request_id, decision, safe_apply_scope);
 
     if !runtime.replace_trace(trace.clone()) {
         return Err(format!("NoName trace not found: {}", trace_id));
@@ -4475,7 +4179,7 @@ pub async fn resolve_noname_second_guardrail(
         review.safe_apply_scope
     };
 
-    record_noname_second_guardrail_decision(&mut trace, &request_id, decision, safe_apply_scope)?;
+    record_second_guardrail_decision(&mut trace, &request_id, decision, safe_apply_scope)?;
 
     if !runtime.replace_trace(trace.clone()) {
         return Err(format!("NoName trace not found: {}", trace_id));
